Fix warning for uintptr_t variance between 32 and 64 bit

+++ test/lib/driver/inc/driver/reg.h
@@ -2,5 +2,5 @@
 
 #include <stdint.h>
 
-#define reg32(base, offset) *(volatile uint32_t*)((base)+(offset))
-#define reg64(base, offset) *(volatile uint64_t*)((base)+(offset))
+#define reg32(base, offset) *(volatile uint32_t*)(uintptr_t)((base)+(offset))
+#define reg64(base, offset) *(volatile uint64_t*)(uintptr_t)((base)+(offset))
